ocr_service/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
pika>=1.3.2
python-dotenv>=1.0.0
pytesseract>=0.3.10
Pillow>=10.1.0
anyio>=3.7.1
pydantic>=2.4.2
starlette>=0.27.0
typing-extensions>=4.7.1
easyocr>=1.7.2
numpy>=1.24.0
opencv-python-headless>=4.8.0
torch>=2.1.0
torchvision>=0.16.0
PyJWT>=2.8.0

